test(inventory-table):Fix test for correct id in product table row
Fix test which looks to see if the product table row has been
created with the correct id. Test was returning the wrong
row as the product table row is the second table row and
querySelector() returns the first matching element. Added
the name property to the table row with a value of productRow
so that querySelector() can return the correct element.